move tic-tac-toe game.js inside tic-tac-toe dir
