+++ examples/modules/tic-tac-toe/components/singleplayer.js
@@ -8,7 +8,7 @@
 
 import React from 'react';
 import Client from 'boardgame.io/client';
-import TicTacToe from '../../../game';
+import TicTacToe from '../game';
 import Board from './board';
 
 const App = Client({
